requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages


pandas
bs4
apache-airflow-providers-postgres==5.10.0